azure-function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
aspect-based-sentiment-analysis==2.0.2
pyodbc==4.0.30